fix: stop defeating the GPU weight cache with a premature .float() call
vulkan_ops._WeightCache (vulkan_ops.py) keys cached GPU weight uploads
on id(weight.untyped_storage()), on the assumption that the same
nn.Module's persistent .weight Parameter object is passed in on every
forward call, so a (potentially large) weight matrix only needs to be
uploaded to the GPU once, not on every single token.

model_runner.py's _wrap_linear (the vLLM Linear-module dispatch hook -
the actual "vllm serve"-integrated path this plugin documents as its
primary usage, distinct from the standalone VulkanModel server) called
vulkan_ops.linear(x.float(), weight.float(), matmul_bias) - converting
weight with .float() *before* passing it in. Tensor.float() allocates a
brand-new tensor with brand-new storage whenever the source isn't
already float32 (true for any bf16/fp16-loaded model, the common
case), so the cache key was different on every single call, and every
call re-did the full host->device weight upload instead of reusing the
cached one - for every wrapped Linear module (q/k/v/o_proj, gate/up/
down_proj, ...) in every decoder layer, on every single token.

vulkan_ops.py's own _vulkan_matvec compounded this: even on a
cache hit, it unconditionally recomputed weight.float() right after
_get_or_upload_weight - a second full weight-matrix copy used only for
.shape[0] (dtype-independent, didn't need the conversion at all) and a
fallback byte-encoding path that's only reachable when the GPU upload
itself failed.

Measured on this hardware, for a single Linear call at a realistic
Gemma4-E2B FFN shape (k=1536, n=6144, bf16 weight): ~13.1ms/call with
the cache defeated vs ~0.94ms/call with it working - a ~13.9x speedup,
compounding across every Linear module (roughly 7 per layer) in every
decoder layer (26-34 for typical models), on every token.

Fixed:
- model_runner.py: _wrap_linear now passes the original weight object
  through unconverted, preserving its identity for the cache.
- vulkan_ops.py: _vulkan_matvec's weight.float() is now lazy (only
  computed in the w_gpu-is-None fallback branch), and N is read from
  weight.shape[0] directly (shape doesn't depend on dtype).

Added tests/python/test_vulkan_ops.py (3 new tests, using a real
Vulkan device, skips cleanly without one): baseline cache-hit sanity
check, a vulkan_ops.linear()-level regression test, and an end-to-end
_wrap_linear regression test - the one that actually reproduces the
bug (the other two don't go through the buggy call site and pass
either way). All three count real uploads via a _WeightCache.put
wrapper rather than inspecting final cache size, since the weak-ref
keyed cache's dict size can - misleadingly - stay stable even when
every call is a fresh miss, if CPython happens to reuse a freed
tensor's memory address for the next allocation. Verified the
regression test fails (6 uploads for 6 calls) on the pre-fix code and
passes (1 upload) after. Verified output correctness is unaffected
(l2 rel err 0.0 vs a torch.nn.functional.linear reference).

ruff/mypy clean; all 32 Python tests pass (2 skipped, no vllm
installed); no Rust changes, so all 19 lib tests continue to pass
unaffected.

+++ vllm_vulkan/model_runner.py
@@ -264,7 +264,18 @@ def _wrap_linear(module: nn.Module) -> None:
         row_reduce = getattr(module, "reduce_results", False) and tp_size > 1
         matmul_bias = None if (row_reduce or skip_bias) else bias
         try:
-            result = vulkan_ops.linear(x.float(), weight.float(), matmul_bias)
+            # weight is passed through as-is (NOT weight.float()) so its
+            # tensor identity - and hence vulkan_ops's persistent GPU
+            # weight cache, keyed on id(weight.untyped_storage()) - stays
+            # stable across every call. weight.float() would allocate a
+            # brand-new tensor (and storage) whenever weight isn't already
+            # float32 (true for any bf16/fp16-loaded model, the common
+            # case), silently defeating the cache and forcing a full
+            # weight-matrix re-upload to the GPU on every single forward
+            # call instead of once. vulkan_ops.linear()/_vulkan_matvec()
+            # do their own float32 conversion internally, lazily, only on
+            # a cache miss (see _get_or_upload_weight).
+            result = vulkan_ops.linear(x.float(), weight, matmul_bias)
             result = result.to(x.dtype)
         except Exception as exc:
             logger.debug("Vulkan linear failed (%s)", exc)
